separate sidebar into component
